==> notebooks/1-EDA-Cleaning.ipynb <==
# Cleaning the dataset from Epicurious.com for SmartRecipes (a recipe recommender)

# Table of Contents
- [Cleaning the dataset from Epicurious.com for the Recipe Recommendation System](#cleaning-the-dataset-from-epicurious.com-for-the-recipe-recommendation-system)
  - [Introduction](#introduction)
  - [Load the data](#load-the-data)
  - [Data Dictionary](#data-dictionary)
  - [Drop columns](#drop-columns)
  - [Convert data types](#convert-data-types)
  - [Check for null values](#check-for-null-values)
  - [Check for duplicates](#check-for-duplicates)
  - [Save the cleaned dataset](#save-the-cleaned-dataset)


## Introduction

We have decided to proceed with the dataset at https://www.kaggle.com/datasets/hugodarwood/epirecipes.

The file of interest to us is `full_format_recipes.json`. 

In this notebook, we will clean the dataset:
- remove columns that are not relevant to us
- remove null and duplicate values

Let's start by loading the dataset.

## Load the data

# import relevant libraries
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns

df_json = pd.read_json('../data/raw/full_format_recipes.json')
df_json.shape

The dataset has 20130 rows and 11 columns

df_json.head()

## Data Dictionary

| Column      | Non-Null Count | Dtype               | Description                                     |
|-------------|----------------|---------------------|-------------------------------------------------|
| ------      | -------------- | -----               | ------                                          |
| title       | 15969 non-null | object              | Title of the recipe                             |
| directions  | 15969 non-null | object              | Steps for the recipe                            |
| ingredients | 15969 non-null | object              | Ingredients plus description of how to cut them |
| categories  | 15969 non-null | object              | Array of categories                             |
| calories    | 15969 non-null | float64             | Calories                                        |
| rating      | 15969 non-null | float64             | Rating on a scale of 0 to 5                     |
| desc        | 10636 non-null | object              | Extra tidbits about the recipe etc.             |
| date        | 15969 non-null | datetime64[ns, UTC] | Date the recipe was created                     |
| sodium      | 15967 non-null | float64             | Sodium content                                  |
| fat         | 15901 non-null | float64             | Fat content                                     |
| protein     | 15922 non-null | float64             | Protein content                                 |

## Drop columns

Columns that are relevant to us:

`title`, `directions`, `ingredients`, `categories`, `calories`, `rating`

We will drop the other columns

cols_to_drop = ["desc",
                "date",
                "sodium",
                "fat",
                "protein"]

df_small = df_json.drop(columns=cols_to_drop)

df_small.shape

After dropping columns, we have 20130 rows and 6 columns.

df_small.info()

df_small.sample(5)

## Convert data types

Some of the columns have lists in them, in order to use the CountVectorizer during preprocessing, we need these to be strings. 

We will convert the contents of the following columns from list to string:
- `directions`
- `categories`
- `ingredients`

print(f"Directions contains entries of type: {type(df_small['directions'][0])}")
print(f"Categories contains entries of type: {type(df_small['categories'][0])}")
print(f"Title contains entries of type: {type(df_small['title'][0])}")
print(f"Ingredients contains entries of type: {type(df_small['ingredients'][0])}")

Let's convert these list types into strings

df_small['directionsStr'] = df_small['directions'].astype(str)
df_small['categoriesStr'] = df_small['categories'].astype(str)
df_small['ingredientsStr'] = df_small['ingredients'].astype(str)


print(f"Directions contains entries of type: {type(df_small['directionsStr'][0])}")
print(f"Categories contains entries of type: {type(df_small['categoriesStr'][0])}")
print(f"Title contains entries of type: {type(df_small['title'][0])}")
print(f"Ingredients contains entries of type: {type(df_small['ingredientsStr'][0])}")

We also noticed an extra space at the end of the contents of `title`. Let's remove that.

df_small['title'][0]

# Remove extra space from the end of the titles
df_small['title'] = df_small['title'].str.strip()

# verify that the space was removed
df_small['title'][0]

Now that we have verified that the conversions look good, we will drop the original columns.

# drop the original columns that contained list types
df_small.drop(columns=['directions','categories','ingredients'], inplace=True)

df_small.info()

## Check for null values

df_small.isna().sum().sum()

There are 4203 missing values.

df_small.isna().sum()

Let's drop these rows that have null values

df_small = df_small.dropna(axis=0)

df_small.isna().sum()

df_small.shape

After dropping rows that have null values, we have 15969 rows and 6 columns.

## Check for duplicates

df_small.duplicated().sum()

There are 1443 duplicates, we will drop these rows.

df_small.drop_duplicates(keep='first', inplace=True)

df_small.duplicated().sum()

df_small.shape

After dropping duplicates, we have 14526 rows and 6 columns

df_small.sample(5)

## Save the cleaned dataset

Let's save the cleaned dataset to a file called full_recipes_cleaned_2.csv

df_small.to_csv("../data/interim/full_recipes_cleaned_2.csv",index=True,index_label='recipeId')
