==> notebooks/2-Pre-processing.ipynb <==
# Pre-processing for the dataset from Epicurious.com for SmartRecipes (a recipe recommender)

# Table of Contents
- [Pre-processing for the dataset from Epicurious.com for the Recipe Recommendation System](#pre-processing-for-the-dataset-from-epicurious.com-for-the-recipe-recommendation-system)
  - [Introduction](#introduction)
  - [Load the cleaned data](#load-the-cleaned-data)
  - [Vectorize](#vectorize)
    - [Define a custom tokenizer](#define-a-custom-tokenizer)
    - [Vectorize ingredientsStr using the custom tokenizer](#vectorize-ingredientsstr-using-the-custom-tokenizer)
  - [Just for fun: Run Logistic Regression to see if we can predict high ratings based on ingredients.](#just-for-fun:-run-logistic-regression-to-see-if-we-can-predict-high-ratings-based-on-ingredients.)
  - [Initial Modeling](#initial-modeling)
    - [Method 1 : Similarity matrix](#method-1-:-similarity-matrix)
    - [Method 2 : Using NearestNeighbors](#method-2-:-using-nearestneighbors)
      - [Matching on "Exact Recipe Title" as Input](#matching-on-"exact-recipe-title"-as-input)
      - [Matching on a "list of ingredients" as Input](#matching-on-a-"list-of-ingredients"-as-input)
      - [Matching on "exclude list of ingredients" as Input](#matching-on-"exclude-list-of-ingredients"-as-input)
      - [Matching on 'include ingredients' + 'exclude ingredients' as input](#matching-on-'include-ingredients'-+-'exclude-ingredients'-as-input)
      - [Trying a vectorizer with a stemmer](#trying-a-vectorizer-with-a-stemmer)
    - [Trying NearestNeighbors on categoriesStr](#trying-nearestneighbors-on-categoriesstr)
      - [Vectorize categoriesStr using the default tokenizer](#vectorize-categoriesstr-using-the-default-tokenizer)
      - [Run NearestNeigbor](#run-nearestneigbor)
  - [For Streamlit](#for-streamlit)
  - [Summary](#summary)
  - [Next Steps and Questions](#next-steps-and-questions)


## Introduction

We have chosen the dataset at https://www.kaggle.com/datasets/hugodarwood/epirecipes?select=full_format_recipes.json for our project.  

We have cleaned this dataset and saved it as `full_recipes_cleaned_2.csv`.  

In this notebook, we will do preprocessing on the data to get it ready for modeling. Since we are dealing with text data, we will use CountVectorizer to preprocess.

We will also do some preliminary modeling.

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import statsmodels.api as sm

from sklearn.feature_extraction.text import CountVectorizer

# custom tokenizer (moved to a separate module due to Streamlit requirements)
import cust_tokenizer


f"pandas version: {pd.__version__}"

## Load the cleaned data

df = pd.read_csv('../data/interim/full_recipes_cleaned_2.csv')
df.shape

df.head()

df.set_index('recipeId')

We have 14526 rows and 6 columns

# confirm that there are no null values or duplicated values
print(f"Null values: {df.isna().sum().sum()}")
print(f"Duplicated rows: {df.duplicated().sum()}")

Verified that there are no null or duplicate values

## Vectorize

### Define a custom tokenizer

<font color='yellow'>NOTE - MOVED my_tokenizer to a module called cust_tokenizer, because of Streamlit</font>

### Vectorize ingredientsStr using the custom tokenizer

from sklearn.feature_extraction.text import CountVectorizer

vectorizer = CountVectorizer(tokenizer=(cust_tokenizer.my_tokenizer),
                       min_df=5)
ingredients_matrix = vectorizer.fit_transform(df['ingredientsStr'])

# IGNORE COMMENTED CODE BELOW

#bigram_series = pd.Series(["chicken breast", "boneless chicken", "chicken stock",
#                           "egg yolks", "chicken thighs", "cocoa powder",
#                           "cilantro leaves", "boneless pork", "sundried tomatoes"])

# join this with the ingredientsStr series
# custom_vocab = pd.concat([df['ingredientsStr'],bigram_series], ignore_index=True)

# specifying bigrams made matters worse, so I will add only the most sensible bigrams manually
# I inspected the results using the code snippet labeled - checking word / bigram frequency

# ingredients_matrix = vectorizer.fit_transform(custom_vocab)
# vectorizer.fit(df['ingredientsStr'])
# ingredients_matrix = vectorizer.transform(df['ingredientsStr'])

# checking word / bigram frequency - used when ngrams was specified in the CountVectorizer
# sum_words = ingredients_matrix.sum(axis = 0)
# words_freq = [(word, sum_words[0, i]) for word, i in vectorizer.vocabulary_.items()]
# words_freq = sorted(words_freq, key = lambda x: x[1], reverse = True)
# words_freq = sorted(words_freq, key = lambda x: x[1])
# words_freq[1000:1500]

ingredients_matrix.shape

ingredients_matrix

# 'Artichoke and Parmesan Risotto' and 'Chicken Parmesan' are two recipes we will use to test for now
df[df['title'].str.contains('Chicken Parmesan', na=False)]

df[df['title'].str.contains('Chicken', na=False)]

## Just for fun: Run Logistic Regression to see if we can predict high ratings based on ingredients.

from sklearn.linear_model import LogisticRegression
from sklearn.model_selection import train_test_split

# Data setup
X = df["ingredientsStr"]
y = df["rating"] >= 4 # binarize our target variable
X_train,X_test,y_train,y_test = train_test_split(X,y,test_size=0.3,
                                                 stratify=y,random_state=42)

# Vectorization
vect_log = CountVectorizer(tokenizer=cust_tokenizer.my_tokenizer,
                       min_df=5)
X_train_vect = vect_log.fit_transform(X_train)
X_test_vect = vect_log.transform(X_test)



# Classifier Training with proper cross validation
best_model = None
best_score = -np.inf
for cval in np.logspace(-3,3,10):
    lr = LogisticRegression(C=cval,max_iter=1000)
    lr.fit(X_train_vect,y_train)
    score = lr.score(X_test_vect,y_test)
    print("Accuracy = {:.4f} for C={:.4f}".format(score,cval))
    if score > best_score:
        best_score = score
        best_model = lr

# finally, evaluate on the test data after fitting with the best hyperparams
best_model.fit(X_train_vect,y_train)
test_acc = best_model.score(X_test_vect,y_test)
print("Our classifier test accuracy is {:.4f}".format(test_acc))

Based on the ingredients only, we can predict a high rating > 4.0 with only 59.66% accuracy. This is slightly better than the baseline, which was 54%. (Earlier we saw that 54% of the dataset has ratings > 4.0.)

## Initial Modeling

### Method 1 : Similarity matrix



<font color="yellow">NOTE: We will use NearestNeighbors, this section of the notebook can be skipped.</font>

Got these steps from the Notebook for the upcoming Recommendation Systems class.

ingredients_matrix[(df['title'] == 'Chicken Parmesan').values].todense().squeeze()

from sklearn.metrics.pairwise import cosine_similarity

recipe_1 = ingredients_matrix[(df['title'] == 'Chicken Parmesan').values,]
recipe_2 = ingredients_matrix[(df['title'] == 'Chicken Parmesan Heros').values,]

print("Similarity:", cosine_similarity(recipe_1, recipe_2)) # Notice the result is a 2D 1X1 array, so to grab
                                                          # the number we will need to index

Using cosine_similarity, 'Chicken Parmesan' and 'Chicken Parmesan Heros' recipes are 34% similar.

recipe_3 = ingredients_matrix[(df['title'] == 'Artichoke and Parmesan Risotto').values,]
print("Similarity:", cosine_similarity(recipe_1, recipe_3))

Using cosine_similarity, 'Chicken Parmesan' and 'Artichoke and Parmesan Risotto' recipes are 20% similar.

Looking at the actual recipes in the dataset this output makes sense.

Let's create a similarity matrix by doing cosine_similarity on the entire ingredients sparse matrix.

from sklearn.metrics.pairwise import cosine_similarity
similarities = cosine_similarity(ingredients_matrix, dense_output=False)

# Check the shape
# rows and columns should be equal, and the number of movies we started with (rows)
similarities.shape

# Test with a sample recipe
df[df['title'] == 'Chicken Parmesan']

# Get the column based upon the index
recipe_index = df[df['title'] == 'Chicken Parmesan'].index

# Create a dataframe with the movie titles
sim_df = pd.DataFrame({'recipe': df['title'],
                       'similarity': np.array(similarities[recipe_index, :].todense()).squeeze()})

# Return the top 10 most similar recipes
sim_df.sort_values(by='similarity', ascending=False).head(10)

With TFIDF Vectorizer:

| recipe |                                       similarity |          |
|-------:|-------------------------------------------------:|----------|
|   4918 |                                 Chicken Parmesan | 1.000000 |
|  11718 |                           Fried Chicken Biscuits | 0.424981 |
|  12825 |                     Quick Baked Chicken Parmesan | 0.424046 |
|   1165 |                          Mozzarella Pesto Spread | 0.412968 |
|   9877 |         Rigatoni with Cheese and Italian Sausage | 0.360720 |
|  14180 |                      BA's Best Eggplant Parmesan | 0.356189 |
|   5786 |                         Horseradish-Yogurt Sauce | 0.351933 |
|   3927 | Chunky Two-Cheese Potatoes with Garlic and Pesto | 0.333565 |
|  13533 |           Fusilli with Shrimp and Paneed Chicken | 0.331478 |
|   3123 |    Spicy Lamb Pizza With Parsley–Red Onion Salad | 0.328433 |

Results with CountVectorizer seem better:

|       |                                            recipe | similarity |
|------:|--------------------------------------------------:|------------|
|  4918 |                                  Chicken Parmesan |   1.000000 |
| 13533 |            Fusilli with Shrimp and Paneed Chicken |   0.587137 |
| 12480 |                          Spicy Oven-Fried Chicken |   0.574641 |
| 11718 |                            Fried Chicken Biscuits |   0.559149 |
| 14180 |                       BA's Best Eggplant Parmesan |   0.557207 |
|  7749 |            Lamb and Eggplant Casserole (Moussaka) |   0.556499 |
| 11245 |                                             Pinon |   0.548443 |
|  9785 | Breaded Skinless Fish Fillets with Red Pepper ... |   0.540222 |
|  6860 |              Chicken and Dumplings with Mushrooms |   0.539425 |
| 11336 | Crispy Chicken Sandwich with Buttermilk Slaw a... |   0.539164 |

TODO: Ana check these results - something does not make sense, should I try with CountVectorizer? Still getting some weird results like Pinon.

Tried with custom tokenizer that removed most measurements and adjectives - TODO see if there is a library for this e.g. https://stackoverflow.com/questions/33587667/extracting-all-nouns-from-a-text-file-using-nltk

The results are better now:
|   |       |                                 recipe | similarity |
|---|------:|---------------------------------------:|------------|
|   |  4918 |                       Chicken Parmesan |   1.000000 |
|   | 12825 |           Quick Baked Chicken Parmesan |   0.553191 |
|   |  6860 |   Chicken and Dumplings with Mushrooms |   0.522931 |
|   | 11718 |                 Fried Chicken Biscuits |   0.520939 |
|   | 13533 | Fusilli with Shrimp and Paneed Chicken |   0.517769 |
|   |  1937 |            East-West Barbecued Chicken |   0.506168 |
|   |  8877 |     Pepper, Rosemary, and Cheese Bread |   0.505992 |
|   |  7471 | Fried Chicken Thighs with Cheesy Grits |   0.503038 |
|   |  4490 |                       Parmesan Muffins |   0.501956 |
|   | 14180 |            BA's Best Eggplant Parmesan |   0.500428 |

selector = (df['title'] == 'Chicken Parmesan') | (df['title'] == 'Fried Chicken Biscuits') | (df['title'] == 'Pinon')
df.loc[selector,['title', 'ingredientsStr']].values

# Check for Aritchoke
# Get the column based upon the index
recipe_index_2 = df[df['title'] == 'Artichoke and Parmesan Risotto'].index

# Create a dataframe with the movie titles
sim_df_2 = pd.DataFrame({'recipe': df['title'],
                       'similarity': np.array(similarities[recipe_index_2, :].todense()).squeeze()})

# Return the top 10 most similar recipes
sim_df_2.sort_values(by='similarity', ascending=False).head(10)

Results look satisfactory.

### Method 2 : Using NearestNeighbors

# Adding a function to print the list of ingredients given a recipe title
def printIngredients(recipeName):
    sel = df['title'] == recipeName
    print(recipeName)
    print(df.loc[sel, ['ingredientsStr']].values)


def printRecipeSteps(recipeName):
    sel = df['title'] == recipeName
    print(recipeName)
    print(df.loc[sel, ['directionsStr']].values)

def printCategories(recipeName):
    sel = df['title'] == recipeName
    print(recipeName)
    print(df.loc[sel, ['categoriesStr']].values)

Our first goal is to find the most similar recipes to a given recipe name that already exists the dataset. For this we will use the NearestNeighbors unsupervised learner.

We will set the following hyperparameters:
- n_neighbors = 11, to find the 10 most similar recipes.
- metric = cosine, we want to find the cosine distance between the recipes.

from sklearn.neighbors import NearestNeighbors

model = NearestNeighbors(n_neighbors=11, metric='cosine')
model.fit(ingredients_matrix)


#### Matching on "Exact Recipe Title" as Input

recipe_index = df[df['title'] == 'Chicken Parmesan'].index

# Run the text input of ingredients through the tokenizer and pass the params into the function below
distances, indices = model.kneighbors(ingredients_matrix[recipe_index])

# recipe_title = df.loc[4918, ['title']]
recipe_titles = []
for id in indices[0]:
    recipe_titles.append(df.loc[id, ['title']])
print(recipe_titles)

recipe_index_2 = df[df['title'] == 'Artichoke and Parmesan Risotto'].index

distances2, indices2 = model.kneighbors(ingredients_matrix[recipe_index_2])
recipe_titles2 = []
for id2 in indices2[0]:
    recipe_titles2.append(df.loc[id2, ['title']])
print(recipe_titles2)

Results seem to be going in the right direction.

#### Matching on a "list of ingredients" as Input

# Using various ingredient lists to test the results

ingInputList = [
    "Chicken, Parmesan, Breadcrumbs",  # something familiar
    # "Artichoke Pesto",
    # "Chicken thighs, potatoes",  # compare results of potatoes vs potato
    # "Chicken thighs, potato",
    "Okra",  # single ingredient
    # "Bhindi",  # unknown ingredient - does not exist in the vocabulary
    # "Peanut butter",  # This is a 2 word ingredient
    # "Eggs, Egg, Eggs, Egg"
]

for ingInput in ingInputList:
    print(f"\n Input ingredients: {ingInput}")
    # Convert the string to a series
    ingInputSeries = pd.Series(ingInput)

    # Let's try to use the vectorizer on this
    ingTransformed = vectorizer.transform(ingInputSeries)

    # pass this to NearestNeighbors trained model
    distOfRes, indicesOfRes = model.kneighbors(ingTransformed)

    # print the output
    print("\n Result")

    for i in range(0, 11):  # TODO: 11 should be made configurable and match the n-neighbors number
        name = df.loc[indicesOfRes[0][i], ['title']].values[0]
        distance = (distOfRes[0][i]).round(3)
        rating = df.loc[indicesOfRes[0][i], ['rating']].values[0]

        # print(f"{name}  :  {distance}")
        print(f"{name}  :  {distance}  :  {rating}")
        printCategories(name)


printIngredients("Onion Oil")
printIngredients("Red-Chile Oil")
printIngredients("Beef Pot Stickers")

printIngredients("Red-Chile Oil")

Stored the results in https://docs.google.com/spreadsheets/d/1hkekdJCZBJqC5hkHKQFFM8sBONUf3CIsv9ekPxqa1AQ/edit?gid=0#gid=0 for easier readability

Results are in the right direction, but:  
- we need to do more tokenization or handle plurals in some way.
- maybe add phrases like 'chicken thighs' 'egg yolks' to tokenizer in some way..

#### Matching on "exclude list of ingredients" as Input

ingInputList = [
    "Chicken, Parmesan, Breadcrumbs",  # something familiar
]

for ingInput in ingInputList:
    print(f"\n Input ingredients: {ingInput}")
    # Convert the string to a series
    ingInputSeries = pd.Series(ingInput)

    # Let's try to use the vectorizer on this
    ingTransformed = (vectorizer.transform(ingInputSeries)) * -1


    # pass this to NearestNeighbors trained model
    distOfRes, indicesOfRes = model.kneighbors(ingTransformed)

    # print the output
    print("\n Result")

    for i in range(0, 11):  # TODO: 11 should be made configurable and match the n-neighbors number
        name = df.loc[indicesOfRes[0][i], ['title']].values[0]
        distance = (distOfRes[0][i]).round(3)
        rating = df.loc[indicesOfRes[0][i], ['rating']].values[0]

        # print(f"{name}  :  {distance}")
        print(f"{name}  :  {distance}  :  {rating}")

    # print ingredient lists to verify
    # for i in range(0, 11):  # TODO: 11 should be made configurable and match the n-neighbors number
    #     name = df.loc[indicesOfRes[0][i], ['title']].values[0]
    #     printIngredients(name)

This worked but one recipe had bread crumbs instead of breadcrumbs and that made it through in the output.

#### Matching on 'include ingredients' + 'exclude ingredients' as input

yes_ing_series = pd.Series("Eggs Egg")
no_ing_series = pd.Series("Spinach")
# NOTE RESULTS FOR THIS COMBO ARE VERY WEIRD

yes_ing_tx = vectorizer.transform(yes_ing_series)
no_ing_tx = (vectorizer.transform(no_ing_series)) * -1  # try increasing to -10 and see what happens

updated_ing_tx = yes_ing_tx + no_ing_tx

distOfRes, indicesOfRes = model.kneighbors(updated_ing_tx)

# print the output
print("\n Result")

for i in range(0, 11):  # TODO: 11 should be made configurable and match the n-neighbors number
    name = df.loc[indicesOfRes[0][i], ['title']].values[0]
    distance = (distOfRes[0][i]).round(3)
    rating = df.loc[indicesOfRes[0][i], ['rating']].values[0]

    # print(f"{name}  :  {distance}")
    print(f"{name}  :  {distance}  :  {rating}")

Seems to be working, although I am concerned about the distance values (it keeps getting larger as I exclude ingredients.)  
Results stored in https://docs.google.com/spreadsheets/d/1hkekdJCZBJqC5hkHKQFFM8sBONUf3CIsv9ekPxqa1AQ/edit?gid=0#gid=0

printIngredients("Chicken Soup")
printRecipeSteps("Chicken Soup")

#### Trying a vectorizer with a stemmer

import nltk
from sklearn.feature_extraction._stop_words import ENGLISH_STOP_WORDS

# Remove units of measurements such as teaspoons, cups, ounces etc. Full list at https://en.wikibooks.org/wiki/Cookbook:Units_of_measurement
measurements = set(line.strip() for line in open('../data/interim/measurement_list.txt'))

# Remove extra adjectives like 'baked', 'thawed', 'cleaned' etc.
extra_adjectives = set(line.strip() for line in open('../data/interim/extra_adjectives_list.txt'))

# Remove some extra words like 'assorted', 'approximately' etc. QUESTION: Is there a smart way to remove the top 100 such words?
extra_words = set(line.strip() for line in open('../data/interim/extra_words_list.txt'))

my_stops = set(ENGLISH_STOP_WORDS) | measurements | extra_adjectives | extra_words
stemmer = nltk.stem.PorterStemmer()

# from 0531_Text_Data
def my_tokenizer_with_stemmer(text):

    # convert to lowercase
    text = text.lower()
    # break into characters and weed out punctuation etc.  (include space!)
    chars = list(char for char in text if char in "abcdefghijklmnopqrstuvwxyz ")
    # make back into a single string
    text = "".join(chars)

    # split sentence into words
    listofwords = text.split(' ')
    listofstemmed_words = []

    # remove stopwords and any tokens that are just empty strings
    for word in listofwords:
        if (not word in ENGLISH_STOP_WORDS) and (word!=''):
            # Stem words
            stemmed_word = stemmer.stem(word)
            listofstemmed_words.append(stemmed_word)

    return listofstemmed_words

vect_stemmed = CountVectorizer(tokenizer=my_tokenizer_with_stemmer,
                       min_df=5)
ingredients_matrix_stemmed = vect_stemmed.fit_transform(df['ingredientsStr'])

# THIS IS 12 seconds slower than the non-stemmed tokenizer :-(

from sklearn.neighbors import NearestNeighbors

model_stemmed = NearestNeighbors(n_neighbors=11, metric='cosine')
model_stemmed.fit(ingredients_matrix_stemmed)


yes_ing_series = pd.Series("Okra")
no_ing_series = pd.Series("Tomato Cilantro")

yes_ing_tx = vect_stemmed.transform(yes_ing_series)
no_ing_tx = (vect_stemmed.transform(no_ing_series)) * -1

updated_ing_tx = yes_ing_tx + no_ing_tx

distOfRes, indicesOfRes = model_stemmed.kneighbors(updated_ing_tx)

# print the output
print("\n Result")

for i in range(0, 11):  # TODO: 11 should be made configurable and match the n-neighbors number
    name = df.loc[indicesOfRes[0][i], ['title']].values[0]
    distance = (distOfRes[0][i]).round(3)
    rating = df.loc[indicesOfRes[0][i], ['rating']].values[0]

    # print(f"{name}  :  {distance}")
    print(f"{name}  :  {distance}  :  {rating}")

# THE RESULTS ARE TOTALLY MESSY ! NO NEED TO STEM THIS WAY

The results with the stemmer are very messy. Not good.

### Trying NearestNeighbors on categoriesStr

This section is a work in progress to see whether we can use the categoriesStr for anything. I have not found anything that works properly so far.

#### Vectorize categoriesStr using the default tokenizer

# NOTE Using default tokenizer here
vect2 = CountVectorizer(stop_words="english")
vect2.fit(df['categoriesStr'])
vocab_categories = vect2.get_feature_names_out()
categories_matrix = vect2.transform(df['categoriesStr'])

len(vocab_categories)

vocab_categories

type(vect2.vocabulary_)

categories_matrix.shape

# Most common categories

sum_cat_words = categories_matrix.sum(axis = 0)
words_freq_cat = [(word, sum_cat_words[0, i]) for word, i in vect2.vocabulary_.items()]
words_freq_cat = sorted(words_freq_cat, key = lambda x: x[1], reverse = True)
words_freq_cat[0:100]

#### Run NearestNeigbor

model_cat = NearestNeighbors(n_neighbors=11, metric='cosine')
model_cat.fit(categories_matrix)
distances_cat1, indices_cat1 = model_cat.kneighbors(categories_matrix[recipe_index])
recipe_cat_titles1 = []
for idc1 in indices_cat1[0]:
    recipe_cat_titles1.append(df.loc[idc1, ['title']])
print(recipe_cat_titles1)


# THIS IS NOT WORKING PROPERLY
catInput = "vegeterian"
catInputSeries = pd.Series(catInput)
catInputTransformed = vect2.transform(catInputSeries)
dcat, icat = model_cat.kneighbors(catInputTransformed)
# print the output
print("\n Result")

for i in range(0, 11):  # TODO: 11 should be made configurable and match the n-neighbors number
    namecat = df.loc[icat[0][i], ['title']].values[0]
    distancecat = (dcat[0][i]).round(3)
    ratingcat = df.loc[icat[0][i], ['categoriesStr']].values[0]

    print(f"{namecat}  :  {distancecat}")
    # print(f"{namecat}  :  {distancecat}  :  {ratingcat}")

distances_cat2, indices_cat2 = model_cat.kneighbors(categories_matrix[recipe_index_2])
recipe_cat_titles2 = []
for idc2 in indices_cat2[0]:
    recipe_cat_titles2.append(df.loc[idc2, ['title']])
print(recipe_cat_titles2)
for recipe in recipe_cat_titles2:
    printCategories(recipe[0])

## Summary

- Just to investigate whether there is a relation between ingredient lists and rating, we ran a Logistic Regression. We found that we can predict a high rating > 4.0 with only 59.66% accuracy. This is slightly better than the baseline, which was 54%. 

- We tried the TFIDF and CountVectorizer with default tokenizer and realized that CountVectorizer suits our purpose better. TFIDF gives higher weights to words that occur less, which is not what we want.

- As a first step to our recommendation system, we wanted to figure out "Give a recipe name, can we find the most similar recipes from the dataset?" We tried 2 ways to do this:
    - Similarity matrix
        - We followed the steps shown in the `Recommendation Systems` notebook. During this process, we found that a custom tokenizer gives us better results, especially when we remove the measurement values (cups, teaspoons etc.) from the vocabulary.
        - We built a custom tokenizer by iteratively comparing results of the similarity matrix.
    - NearestNeighbors
        - We figured out how to use this and saw that the results were the same as the Similarity matrix method but with much less headaches while coding. We will use this method going forward.

- We were able to find similar recipes satisfactorily. A few issues that need further investigation / fixing:
    - handling plurals - potato / potatoes
    - handling 2 word ingredients - "peanut butter"

- Next we updated the code to use a list of ingredients instead of a recipe name. And also included an option to exclude ingredients. More investigation needs to be done on the quality of results. This is a work in progress.

## Next Steps and Questions

vocab = vectorizer.get_feature_names_out()

vocab

custom_list = list(vocab) + ["peanut butter"]
custom_list

Tokenization questions  
- Is there a way to include 2 word ingredients in the vocabulary manually? Example "Peanut butter" This is vastly different from just Peanuts. 
- How to handle plurals? Stemming is slow and didn't really help.
- Interesting article about tokenization - https://www.kaggle.com/code/shivanirana63/beginner-s-guide-to-word-tokenization
- Can I use categories to pick vegetarian recipes too? From categories, how to tokenize ONLY specific ones like 'vegetarian', 'pescatarian', 'wheat', 'gluten' ?

Ranking  
- When distance is the same how to rank the recipe?

General  
- Should I create a column with number of ingredients? May not be as straigforward.
- Do fuzzy matching or spell check on the input where the vocabulary will be from the vectorizer.
