==> notebooks/4-Final-Modelling.ipynb <==
# Model and Evaluation for SmartRecipes (a recipe recommender)

## Table of Contents
- [Model and Evaluation for the SmartRecipes Recipe Recommender](#model-and-evaluation-for-the-smartrecipes-recipe-recommender)
  - [Introduction](#introduction)
  - [Goal](#goal)
  - [Process](#process)
    - [Load the cleaned dataset](#load-the-cleaned-dataset)
    - [Feature Engineering `featuredCol` = `categories` + `ingredients`](#feature-engineering-featuredcol-=-categories-+-ingredients)
    - [Vectorize `ingredientsStr` to get basic vocabulary](#vectorize-ingredientsstr-to-get-basic-vocabulary)
    - [Create custom vocabulary](#create-custom-vocabulary)
    - [Get the transformed sparse matrix by vectorizing `featuredCol`](#get-the-transformed-sparse-matrix-by-vectorizing-featuredcol)
    - [Create the `NearestNeighbors` model](#create-the-nearestneighbors-model)
    - [Test and Evaluate the model](#test-and-evaluate-the-model)
    - [Save for Streamlit](#save-for-streamlit)
  - [Next Steps](#next-steps)


## Introduction

1. So far we have cleaned the dataset
2. Done some experimentation and iteration and figured out a custom tokenizer for or CountVectorizer.
3. Figured out how to match for recipes based on a list of ingredients that the user has entered.
4. Figured out that we will use a custom vocabulary to include 2 word ingredients, since using bigrams in the vectorizer's hyperparameters was not giving us satisfactory results.
4. The results are returned in order of number of ingredients, which matches our reducing food waste goal.
5. Figured out that we get better results if we include the categories string in the vectorization too.
6. Shortlisted some categories that are of interest to us - []"Drink", "Dairy Free", "Gluten free", "Dessert", "Vegetarian"]. 

## Goal

In this notebook our goal is to combine these various findings into a sequential series of steps:

1. Feature engineering - Combine the `categoriesStr` and `ingredientsStr` column. Name it `featuredCol`
2. Vectorize the `ingredientsStr` column using the custom tokenizer. We want most of the words from this column. This step will give us the vocabulary.
3. Create the custom vocabulary `custom_vocab` by adding the two word ingredients like "peanut butter" and any category words of interest like "vegetarian".
4. Get the ingredients + relevant categories sparse matrix by calling transform() on the `featuredCol` using a 2nd CountVectorizer that has the `custom_vocab` and custom tokenizer. Note: We will not call fit() on this 2nd vectorizer because we are passing a custom vocabulary. We will directly call transform().
5. Create a model using NearestNeighbors and fit it using the sparse matrix got in the step above.
6. Use the 2nd vectorizer to transform any input before giving it to the model.
7. Call model.kneighbors() to get the results.
8. For Streamlit usage.
    1. Save the `custom_vocab`
    2. Save the `custom tokenizer`
    3. Save the 2nd vectorizer.
    4. Save the NearestNeighbors model.
    5. Save the dataset with the `featuredCol` 

## Process

Let's start by loading the cleaned dataset.

### Load the cleaned dataset

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
# import seaborn as sns
# import statsmodels.api as sm
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.neighbors import NearestNeighbors

#(NOTE: joblib had to be installed the first time I used it)
import joblib
# custom tokenizer (moved to a separate module due to Streamlit requirements)
import cust_tokenizer


df = pd.read_csv('../data/interim/full_recipes_cleaned_2.csv')
df.shape

df.set_index('recipeId')

# confirm that there are no null values or duplicated values
print(f"Null values: {df.isna().sum().sum()}")
print(f"Duplicated rows: {df.duplicated().sum()}")

### Feature Engineering `featuredCol` = `categories` + `ingredients`

df["featuredCol"] = df['categoriesStr'] + df['ingredientsStr']

type(df.loc[0,['featuredCol']].values[0])

### Vectorize `ingredientsStr` to get basic vocabulary

Vectorize the ingredient list to get the vocab

from sklearn.feature_extraction.text import CountVectorizer

vectorizer = CountVectorizer(tokenizer=(cust_tokenizer.my_tokenizer),
                       min_df=5)
ingredients_matrix = vectorizer.fit_transform(df['ingredientsStr'])


### Create custom vocabulary

Add 2 word ingredients and the categories you want to the custom vocab

new_vocab_list = list(vectorizer.get_feature_names_out())
new_vocab_list.append("Peanut butter")
new_vocab_list.extend(["Drink", "Dairy Free", "Gluten free", "Dessert", "Vegetarian"])
len(new_vocab_list)

### Get the transformed sparse matrix by vectorizing `featuredCol`

Now transform the featuredCol

# Let's set this new vocab as a hyperparameter of the vectorizer
vectorizer_mod = CountVectorizer(tokenizer=(cust_tokenizer.my_tokenizer),
                       min_df=5, vocabulary=new_vocab_list)
# not calling fit because sending custom vocab..
ingredients_matrix_mod = vectorizer_mod.transform(df['featuredCol'])


# Adding a function to print the list of ingredients given a recipe title
def printIngredients(recipeName):
    sel = df['title'] == recipeName
    print(recipeName)
    print(df.loc[sel, ['featuredCol']].values)


### Create the `NearestNeighbors` model

model = NearestNeighbors(n_neighbors=11, metric='cosine')
model.fit(ingredients_matrix_mod)

### Test and Evaluate the model

# Using various ingredient lists to test the results

ingInputList = [
    # "Chicken, Parmesan, Breadcrumbs",  # something familiar
    # "Artichoke Pesto",
    # "Chicken thighs, potatoes",  # compare results of potatoes vs potato
    # "Chicken thighs, potato",
    # "Okra",  # single ingredient
    # "Bhindi",  # unknown ingredient - does not exist in the vocabulary
    "Peanut butter"  # This is a 2 word ingredient
]

for ingInput in ingInputList:
    print(f"\n Input ingredients: {ingInput}")
    # Convert the string to a series
    ingInputSeries = pd.Series(ingInput)

    # Let's try to use the vectorizer on this
    ingTransformed = vectorizer_mod.transform(ingInputSeries)

    # pass this to NearestNeighbors trained model
    distOfRes, indicesOfRes = model.kneighbors(ingTransformed)

    # print the output
    print("\n Result")

    for i in range(0, 11):  # TODO: 11 should be made configurable and match the n-neighbors number
        name = df.loc[indicesOfRes[0][i], ['title']].values[0]
        distance = (distOfRes[0][i]).round(3)
        rating = df.loc[indicesOfRes[0][i], ['rating']].values[0]

        # print(f"{name}  :  {distance}")
        print(f"{name}  :  {distance}  :  {rating}")


ingInputList = [
    # "Chicken, Parmesan, Breadcrumbs",  # something familiar
    # "Artichoke Pesto",
    "Chicken thighs, potatoes",  # compare results of potatoes vs potato
    "Chicken thighs, potatoes, gluten free",  # compare results of potatoes vs potato
    # "Chicken thighs, potato",
    "Okra",  # single ingredient
    "Okra, vegetarian",
    # "Bhindi",  # unknown ingredient - does not exist in the vocabulary
    # "Peanut butter"  # This is a 2 word ingredient
]

for ingInput in ingInputList:
    print(f"\n Input ingredients: {ingInput}")
    # Convert the string to a series
    ingInputSeries = pd.Series(ingInput)

    # Let's try to use the vectorizer on this
    ingTransformed = vectorizer_mod.transform(ingInputSeries)

    # pass this to NearestNeighbors trained model
    distOfRes, indicesOfRes = model.kneighbors(ingTransformed)

    # print the output
    print("\n Result")

    for i in range(0, 11):  # TODO: 11 should be made configurable and match the n-neighbors number
        name = df.loc[indicesOfRes[0][i], ['title']].values[0]
        distance = (distOfRes[0][i]).round(3)
        rating = df.loc[indicesOfRes[0][i], ['rating']].values[0]

        # print(f"{name}  :  {distance}")
        print(f"{name}  :  {distance}  :  {rating}")


Results look satisfactory - Comparison of various runs is in https://docs.google.com/spreadsheets/d/1hkekdJCZBJqC5hkHKQFFM8sBONUf3CIsv9ekPxqa1AQ/edit?gid=0#gid=0

printIngredients("Chicken Vegetable Soup with Lime and Cilantro")

printIngredients("Cane Vinegar Chicken with Pearl Onions, Orange & Spinach")

We can see that specifying "Gluten-free" changes the results.

yes_ing_series = pd.Series("Broccoli")
no_ing_series = pd.Series("Asparagus")
# Results look a little strange for this combo.

yes_ing_tx = vectorizer_mod.transform(yes_ing_series)
no_ing_tx = (vectorizer_mod.transform(no_ing_series)) * -1  # try increasing to -10 and see what happens

updated_ing_tx = yes_ing_tx + no_ing_tx

distOfRes, indicesOfRes = model.kneighbors(updated_ing_tx)

# print the output
print("\n Result")

for i in range(0, 11):  # TODO: 11 should be made configurable and match the n-neighbors number
    name = df.loc[indicesOfRes[0][i], ['title']].values[0]
    distance = (distOfRes[0][i]).round(3)
    rating = df.loc[indicesOfRes[0][i], ['rating']].values[0]

    # print(f"{name}  :  {distance}")
    print(f"{name}  :  {distance}  :  {rating}")

### Save for Streamlit

1. Save the `custom_vocab`
2. Save the `custom tokenizer` (done already as a module cust_tokenizer)
3. Save the 2nd vectorizer.
4. Save the NearestNeighbors model.
5. Save the dataset with the `featuredCol`

# Save the custom vocab
joblib.dump(new_vocab_list, '../model/custom_vocab.pkl')

# Save the 2nd vectorizer.
joblib.dump(vectorizer_mod, "../model/vect_mod.pkl")

# Save the NearestNeighbors model.
joblib.dump(model, "../model/model_final.pkl")

# Save the dataset with the featuredCol
df.to_csv("../data/final/full_recipes.csv",index=True,index_label='recipeId')

## Next Steps

- This was an unsupervised learning problem, so the results had to be evaluated manually.

- Recipes with the least relevant ingredients show up in the top 5 results ~ 80% of the time.

- Build a Streamlit app for the recipe recommender.
